+++ doc/examples/animations/lib/main.dart
@@ -1,9 +1,7 @@
+import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:flame/gestures.dart';
 import 'package:flutter/gestures.dart';
-import 'package:flame/extensions/vector2.dart';
-import 'package:flame/sprite_animation.dart';
-import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart' hide Image;
 import 'dart:ui';
 
@@ -39,11 +37,13 @@ class MyGame extends BaseGame with TapDetector {
     );
 
     final spriteSize = Vector2.all(100.0);
-    final animationComponent2 = SpriteAnimationComponent(spriteSize, animation);
+    final animationComponent2 =
+        SpriteAnimationComponent.fromSpriteAnimation(spriteSize, animation);
     animationComponent2.x = size.x / 2 - spriteSize.x;
     animationComponent2.y = spriteSize.y;
 
-    final reversedAnimationComponent = SpriteAnimationComponent(
+    final reversedAnimationComponent =
+        SpriteAnimationComponent.fromSpriteAnimation(
       spriteSize,
       animation.reversed(),
     );