+++ doc/examples/aseprite/lib/main.dart
@@ -1,7 +1,5 @@
+import 'package:flame/components.dart';
 import 'package:flame/game.dart';
-import 'package:flame/sprite_animation.dart';
-import 'package:flame/extensions/vector2.dart';
-import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
@@ -22,8 +20,9 @@ class MyGame extends BaseGame {
       jsonData,
     );
     final spriteSize = Vector2.all(200);
-    final animationComponent = SpriteAnimationComponent(spriteSize, animation)
-      ..position = size / 2 - Vector2.all(100);
+    final animationComponent =
+        SpriteAnimationComponent.fromSpriteAnimation(spriteSize, animation)
+          ..position = size / 2 - Vector2.all(100);
 
     add(animationComponent);
   }