+++ doc/examples/effects/combined_effects/lib/main.dart
@@ -1,13 +1,9 @@
-import 'package:flame/effects/combined_effect.dart';
-import 'package:flame/effects/move_effect.dart';
-import 'package:flame/effects/scale_effect.dart';
-import 'package:flame/effects/rotate_effect.dart';
+import 'package:flame/effects.dart';
 import 'package:flame/game.dart';
 import 'package:flame/gestures.dart';
-import 'package:flame/extensions/offset.dart';
-import 'package:flame/extensions/vector2.dart';
 import 'package:flame/flame.dart';
-import 'package:flutter/material.dart';
+import 'package:flame/extensions.dart';
+import 'package:flutter/widgets.dart';
 
 import './square.dart';