+++ doc/examples/parallax/lib/main_advanced.dart
@@ -1,7 +1,6 @@
+import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/parallax.dart';
-import 'package:flame/components/parallax_component.dart';
-import 'package:flame/extensions/vector2.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart';