+++ doc/examples/widgets/lib/main.dart
@@ -1,15 +1,9 @@
-import 'package:flame/extensions/vector2.dart';
+import 'package:flame/extensions.dart';
+import 'package:flame/widgets.dart';
 import 'package:flutter/material.dart' hide Animation;
 import 'package:flame/flame.dart';
-import 'package:flame/sprite.dart';
 import 'package:flame/spritesheet.dart';
 import 'package:dashbook/dashbook.dart';
-
-import 'package:flame/widgets/nine_tile_box.dart';
-import 'package:flame/widgets/sprite_button.dart';
-import 'package:flame/widgets/sprite_widget.dart';
-import 'package:flame/widgets/animation_widget.dart';
-import 'package:flame/anchor.dart';
 
 Anchor parseAnchor(String name) {
   switch (name) {
@@ -33,7 +27,7 @@ Anchor parseAnchor(String name) {
       return Anchor.bottomRight;
   }
 
-  return null;
+  throw Exception("Cannot parse anchor name `$name`");
 }
 
 void main() async {
@@ -108,9 +102,8 @@ void main() async {
           child: SpriteWidget(
             sprite: shieldSprite,
             anchor: parseAnchor(
-                  ctx.listProperty('anchor', 'Anchor.center', anchorOptions),
-                ) ??
-                Anchor.topLeft,
+              ctx.listProperty('anchor', 'Anchor.center', anchorOptions),
+            ),
           ),
         ),
       );
@@ -126,7 +119,7 @@ void main() async {
     to: 3,
     loop: true,
   );
-  dashbook.storiesOf('AnimationWidget').decorator(CenterDecorator()).add(
+  dashbook.storiesOf('SpriteAnimationWidget').decorator(CenterDecorator()).add(
         'default',
         (ctx) => Container(
           width: ctx.numberProperty('container width', 400),
@@ -135,9 +128,8 @@ void main() async {
             animation: _animation,
             playing: ctx.boolProperty('playing', true),
             anchor: parseAnchor(
-                  ctx.listProperty('anchor', 'Anchor.center', anchorOptions),
-                ) ??
-                Anchor.topLeft,
+              ctx.listProperty('anchor', 'Anchor.center', anchorOptions),
+            ),
           ),
         ),
       );