+++ lib/components/mixins/has_game_ref.dart
@@ -1,5 +1,0 @@
-import '../../game.dart';
-
-mixin HasGameRef<T extends Game> {
-  T gameRef;
-}