+++ test/components/position_component_test.dart
@@ -1,11 +1,11 @@
 import 'dart:math' as math;
 
-import 'package:flame/anchor.dart';
-import 'package:flame/components/position_component.dart';
-import 'package:flame/extensions/vector2.dart';
+import 'package:flame/components.dart';
 import 'package:test/test.dart';
 
 class MyComponent extends PositionComponent {}
+
+class MyHitboxComponent extends PositionComponent with Hitbox {}
 
 void main() {
   group('PositionComponent overlap test', () {
@@ -17,7 +17,7 @@ void main() {
       component.anchor = Anchor.center;
 
       final point = Vector2(2.0, 2.0);
-      expect(component.checkOverlap(point), true);
+      expect(component.containsPoint(point), true);
     });
 
     test('overlap on edge', () {
@@ -28,7 +28,7 @@ void main() {
       component.anchor = Anchor.center;
 
       final point = Vector2(1.0, 1.0);
-      expect(component.checkOverlap(point), true);
+      expect(component.containsPoint(point), true);
     });
 
     test('not overlapping with x', () {
@@ -39,7 +39,7 @@ void main() {
       component.anchor = Anchor.center;
 
       final point = Vector2(4.0, 1.0);
-      expect(component.checkOverlap(point), false);
+      expect(component.containsPoint(point), false);
     });
 
     test('not overlapping with y', () {
@@ -50,7 +50,7 @@ void main() {
       component.anchor = Anchor.center;
 
       final point = Vector2(1.0, 4.0);
-      expect(component.checkOverlap(point), false);
+      expect(component.containsPoint(point), false);
     });
 
     test('overlapping with angle', () {
@@ -61,7 +61,7 @@ void main() {
       component.anchor = Anchor.center;
 
       final point = Vector2(3.1, 2.0);
-      expect(component.checkOverlap(point), true);
+      expect(component.containsPoint(point), true);
     });
 
     test('not overlapping with angle', () {
@@ -72,7 +72,7 @@ void main() {
       component.anchor = Anchor.center;
 
       final point = Vector2(1.0, 0.1);
-      expect(component.checkOverlap(point), false);
+      expect(component.containsPoint(point), false);
     });
 
     test('overlapping with angle and topLeft anchor', () {
@@ -83,7 +83,41 @@ void main() {
       component.anchor = Anchor.topLeft;
 
       final point = Vector2(1.0, 3.1);
-      expect(component.checkOverlap(point), true);
+      expect(component.containsPoint(point), true);
+    });
+
+    test('component with hitbox contains point', () {
+      final size = Vector2(2.0, 2.0);
+      final Hitbox component = MyHitboxComponent();
+      component.position = Vector2(1.0, 1.0);
+      component.anchor = Anchor.topLeft;
+      component.size = size;
+      component.shape = [
+        Vector2(1, 0),
+        Vector2(0, -1),
+        Vector2(-1, 0),
+        Vector2(0, 1),
+      ];
+
+      final point = component.position + component.size / 4;
+      expect(component.containsPoint(point), true);
+    });
+
+    test('component with hitbox does not contains point', () {
+      final size = Vector2(2.0, 2.0);
+      final Hitbox component = MyHitboxComponent();
+      component.position = Vector2(1.0, 1.0);
+      component.anchor = Anchor.topLeft;
+      component.size = size;
+      component.shape = [
+        Vector2(1, 0),
+        Vector2(0, -1),
+        Vector2(-1, 0),
+        Vector2(0, 1),
+      ];
+
+      final point = Vector2(1.1, 1.1);
+      expect(component.containsPoint(point), false);
     });
   });
 }