+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-## [next]
+## 1.0.0-rc6
+ - Use `Offset` type directly in `JoystickAction.update` calculations
+ - Changed `parseAnchor` in `examples/widgets` to throw an exception instead of returning null when it cannot parse an anchor name
  - Code improvements and preparing APIs to null-safety
  - BaseComponent removes children marked as shouldRemove during update
  - Use `find` instead of `globstar` pattern in `scripts/lint.sh` as the later isn't enabled by default in bash
@@ -11,6 +13,18 @@
  - Adding shortcut for loading Sprites and SpriteAnimation from the global cache
  - Adding loading methods for the different `ParallaxComponent` parts and refactor how the delta velocity works
  - Add tests for `Timer` and fix a bug where `progress` was not reported correctly
+ - Refactored the `SpriteBatch` class to be more elegant and to support `Vector2`.
+ - Added fallback support for the web on the `SpriteBatch` class
+ - Added missing documentation on the `SpriteBatch` class
+ - Added an utility method to load a `SpriteBatch` on the `Game` class
+ - Updated the `widgets.md` documentation
+ - Removing methods `initialDimensions` and `removeGestureRecognizer` to avoid confusion
+ - Adding standard for `SpriteComponent` and `SpriteAnimationComponent` constructors
+ - Added `Composition`, allows for composing multiple images into a single image.
+ - Move files to comply with the dart package layout convention
+ - Fix gesture detection bug of children of `PositionComponent`
+ - The `game` argument on `GameWidget` is now required
+ - Add hitbox mixin for PositionComponent to make more accurate gestures
 
 ## 1.0.0-rc5
  - Option for overlays to be already visible on the GameWidget